% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_2D.R
\name{Bayes_boundary_1D}
\alias{Bayes_boundary_1D}
\title{Mixture model for k in 1:2:}
\usage{
Bayes_boundary_1D(xy_params = NULL, tol = 1e-10)
}
\arguments{
\item{xy_params}{<tbl> prescribed mean, sd, prob, label per group}

\item{tol}{<dbl> lower bound to compare abs(x) to zero}
}
\value{
list solution type, roots, and coefficients
}
\description{
(X | Y = k) = X_k
= mu_k + sigma_k * Z_k
P(Y = k) = p_k
Problem: determine d_x, the Bayes optimal decision boundary
between X_1 and X_2.
Return: list solution type, roots, and coefficients.
}
