% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_2D.R
\name{bbox_diag_to_seg}
\alias{bbox_diag_to_seg}
\title{Given bounding box}
\usage{
bbox_diag_to_seg(bbox = NULL)
}
\arguments{
\item{bbox}{<lst> named list(xmin, xmax, ymin, ymax)}
}
\value{
main diagonal segment of input bounding box
}
\description{
bbox = list(xmin, xmax, ymin, ymax)
return segment
seg = c(x = xmin, y = ymin, xend = xmax, yend = ymax)
}
