% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_2D.R
\name{bbox_valid}
\alias{bbox_valid}
\title{Validate a purported bounding box,}
\usage{
bbox_valid(bbox)
}
\arguments{
\item{bbox}{<lst> named list (xmin, xmax, ymin, ymax)}
}
\value{
validation (T/F) of input bounding box
}
\description{
which is a list of the following form.
bbox = list (xmin, xmax, ymin, ymax)
}
