% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_2D.R
\name{bt_list}
\alias{bt_list}
\title{Find the points of intersection of a bounding box and a line.}
\usage{
bt_list(bbox, trip, tol = 1e-10)
}
\arguments{
\item{bbox}{<lst> named list (xmin, xmax, ymin, ymax)}

\item{trip}{<dbl> triplet (c0, c1, c2)}

\item{tol}{<dbl> lower bound for norm(c1, c2)}
}
\value{
(line, bounding box) points of intersection
}
\description{
bbox: (xmin, xmax, ymin, ymax)
line: c0 + c1 x + c2 y = 0
Return a list of 2 intersection points or else return NULL.
}
