% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_2D.R
\name{bt_seg}
\alias{bt_seg}
\title{Find the 2 points of intersection of a bounding box and a line.}
\usage{
bt_seg(bbox, trip, tol = 1e-10)
}
\arguments{
\item{bbox}{<lst> named list (xmin, xmax, ymin, ymax)}

\item{trip}{<dbl> un-named triplet (c0, c1, c2)}

\item{tol}{<dbl> lower bound for norm(c1, c2)}
}
\value{
segment: (line, bounding box) intersection
}
\description{
bbox: (xmin, xmax, ymin, ymax)
line: c0 + c1 x + c2 y = 0
Return a directed segment in the form of a named
vector (x, y, xend, yend) or else return NULL.
}
