% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_2D.R
\name{extract_ggplot_bbox}
\alias{extract_ggplot_bbox}
\title{Given a ggplot object, extract the actual axis limits}
\usage{
extract_ggplot_bbox(gg_plot)
}
\arguments{
\item{gg_plot}{<lst> object returned by ggplot2::ggplot()}
}
\value{
bounding box for input ggplot object
}
\description{
that will be used for rendering.
}
