% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_circuit.R
\name{gen_circuit_tbl}
\alias{gen_circuit_tbl}
\title{return a tibble of points (c_j, s_j) along}
\usage{
gen_circuit_tbl(n_pts = 200L, f_0 = 1/2, f_1 = 1/4, period = 1/2, phase = 0)
}
\arguments{
\item{n_pts}{<int> desired number of points}

\item{f_0}{<dbl> positive constant}

\item{f_1}{<dbl> amplitude (f_1 < f_0)}

\item{period}{<dbl> period of the sine wave}

\item{phase}{<dbl> phase of the sine wave (mod 1)}
}
\value{
point pairs along simple closed curve
}
\description{
a simple closed curve in the (x, y) plane,
namely the unit circle deformed by norm f(t)
}
