% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_circuit.R
\name{gen_wave_norm}
\alias{gen_wave_norm}
\title{return a sine wave function f(t) > 0}
\usage{
gen_wave_norm(f_0 = 1/2, f_1 = 1/4, period = 1/2, phase = 0)
}
\arguments{
\item{f_0}{<dbl> positive constant}

\item{f_1}{<dbl> amplitude (f_1 < f_0)}

\item{period}{<dbl> period of the sine wave}

\item{phase}{<dbl> phase of the sine wave (mod 1)}
}
\value{
sine wave function f(t) > 0
}
\description{
to serve as the varying norm of points
(c(t), s(t)) along a curve
}
