% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_ds_tbl.R
\name{get_all_ds_per_pkg}
\alias{get_all_ds_per_pkg}
\title{return tibble of all data set names per package}
\usage{
get_all_ds_per_pkg(pkgs)
}
\arguments{
\item{pkgs}{<chr> package name(s)}
}
\value{
data set names per package
}
\description{
note: named packages are assumed to installed
}
