% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ref_materials.R
\name{get_articles_tbl}
\alias{get_articles_tbl}
\title{-- return a tibble of reference articles}
\usage{
get_articles_tbl()
}
\value{
tibble of reference articles
}
\description{
-- return a tibble of reference articles
}
