% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_2D.R
\name{get_bb_segs}
\alias{get_bb_segs}
\title{Given: data frame df, 2 specified predictor columns,}
\usage{
get_bb_segs(df, x_1, x_2, y_group, bbox = NULL)
}
\arguments{
\item{df}{<df>  a data frame containing (x_1, x_2, y_group)}

\item{x_1}{<id>  name of 1st predictor variable}

\item{x_2}{<id>  name of 2nd predictor variable}

\item{y_group}{<id>  name of grouping variable}

\item{bbox}{<lst> named list (xmin, xmax, ymin, ymax)}
}
\value{
segments giving bounding box per group of points
}
\description{
(x_1, x_2), 1 specified grouping column (y_group),
and a bounding box (bbox) for a scatter-plot.
Return: a list of statistical tables culminating in
a table of bounding box segments, one directed segment
for each pair of distinct groups.
}
