% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ref_materials.R
\name{get_books_tbl}
\alias{get_books_tbl}
\title{-- return a tibble of reference books}
\usage{
get_books_tbl()
}
\value{
tibble of reference books
}
\description{
-- return a tibble of reference books
}
