% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_core_packages.R
\name{get_core_packages}
\alias{get_core_packages}
\title{Core packages for EDA for ML student projects}
\usage{
get_core_packages()
}
\value{
Character vector of package names.
}
\description{
Returns the vector of packages loaded in the setup chunk of every
`chNN-work.qmd` generated by [create_project()].
}
\keyword{internal}
