% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_2D.R
\name{get_default_xy_params}
\alias{get_default_xy_params}
\title{Return 2-by-4 tibble: mean, sd, prob, label per group}
\usage{
get_default_xy_params()
}
\value{
(mean, sd, prob, label) tibble per group
}
\description{
default example: book price (general, technical)
url: https://latestcost.com/average-cost-of-hardcover-book/
}
