% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_2D.R
\name{get_fn_trip_line}
\alias{get_fn_trip_line}
\title{Given a triplet, (c0, c1, c2), of homogeneous}
\usage{
get_fn_trip_line(trip)
}
\arguments{
\item{trip}{<dbl> triplet c(c0, c1, c2)}
}
\value{
bi-linear function defined by homogeneous coordinates
}
\description{
coefficients, return the following function
of (x, y):
f(x, y) =  c0 + c1 x + c2 y
}
