% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_2D.R
\name{get_igroup_tbl}
\alias{get_igroup_tbl}
\title{Generate a tibble of integer indices.}
\usage{
get_igroup_tbl(n_groups = 4L)
}
\arguments{
\item{n_groups}{<int> determines group indices 1:n_groups}
}
\value{
tibble of C(n_groups, 2) distinct ordered index pairs
}
\description{
Generate a tibble of integer indices.
}
