% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_2D.R
\name{get_iseg_tbl}
\alias{get_iseg_tbl}
\title{Given igroup_tbl: containing the indices (ig_1, ig_2)}
\usage{
get_iseg_tbl(igroup_tbl)
}
\arguments{
\item{igroup_tbl}{<tbl> starts with columns (ig_1, ig_2, ...)}
}
\value{
pairs of index-pairs that share exactly one index
}
\description{
of pairs of distinct groups having a bbox line segment.
Return iseg_tbl: (is_1, is_2, ig_1, ig_2, ig_3, ig_4),
consisting of the indices of selected pairs of distinct
segments and their corresponding group indices.
Selection criteria:
(1) index pair (ig_1, ig_2) has exactly one integer
value in common with (ig_3, ig_4).
(2) each possible segment combination (is_1, is_2)
occurs no more than once in the returned tibble.
}
