% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_aesthetics.R
\name{get_pc_segs}
\alias{get_pc_segs}
\title{Return segment endpoint coordinates}
\usage{
get_pc_segs(prcomp_obj, mu_max = NULL, mu_sd = NULL)
}
\arguments{
\item{prcomp_obj}{<lst> object returned by stats::prcomp()}

\item{mu_max}{<dbl> multiple of max score}

\item{mu_sd}{<dbl> multiple of sd score}
}
\value{
matrix defining one segment per principle component
}
\description{
for each principal component.
Use mu_sd if specified.
If not, use mu_max (specified or default).
}
