% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_aesthetics.R
\name{get_pc_stats}
\alias{get_pc_stats}
\title{Return the standard deviation and variance}
\usage{
get_pc_stats(prcomp_obj)
}
\arguments{
\item{prcomp_obj}{<lst> object returned by stats::prcomp()}
}
\value{
tibble: (sd, var) of each principle component
}
\description{
of each principal component.
}
