% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_aesthetics.R
\name{get_rot_tbl_lst}
\alias{get_rot_tbl_lst}
\title{Return the rotation matrix as a tibble}
\usage{
get_rot_tbl_lst(prcomp_obj)
}
\arguments{
\item{prcomp_obj}{<lst> object returned by stats::prcomp()}
}
\value{
list: rotation matrix as (long, wide) tibble
}
\description{
in both wide and long configurations.
}
