% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_2D.R
\name{get_ss_points}
\alias{get_ss_points}
\title{Given: data frame df, 2 specified predictor columns,}
\usage{
get_ss_points(df, x_1, x_2, y_group, bbox = NULL)
}
\arguments{
\item{df}{<df>  a data frame containing (x_1, x_2, y_group)}

\item{x_1}{<id>  name of 1st predictor variable}

\item{x_2}{<id>  name of 2nd predictor variable}

\item{y_group}{<id>  name of grouping variable}

\item{bbox}{<lst> named list (xmin, xmax, ymin, ymax)}
}
\value{
list: discriminating segments, their intersections
}
\description{
(x_1, x_2), 1 specified grouping column (y_group),
and a bounding box (bbox) for a scatter-plot.
Return: statistical tables culminating in the tabulation
of pairs of distinct bounding box line segments.  Each
segment distinguishes a pair of distinct groups.  For
each pair of segments a point of intersection is calculated.
}
