% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_per_state.R
\name{get_state_abb_nm}
\alias{get_state_abb_nm}
\title{Return tibble of (abbreviation, name) per state}
\usage{
get_state_abb_nm()
}
\value{
tibble of (abbreviation, name) per state
}
\description{
Return tibble of (abbreviation, name) per state
}
