% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_per_state.R
\name{get_state_ctr}
\alias{get_state_ctr}
\title{Return tibble with geographic center of each US state}
\usage{
get_state_ctr()
}
\value{
tibble with geographic center of each US state
}
\description{
Return tibble with geographic center of each US state
}
