% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_per_state.R
\name{get_state_x77}
\alias{get_state_x77}
\title{Return tibble of datasets::state.x77}
\usage{
get_state_x77()
}
\value{
tibble: statistics per state from datasets::state.x77
}
\description{
Return tibble of datasets::state.x77
}
