% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_2D.R
\name{get_std_trip}
\alias{get_std_trip}
\title{Ensure that a given numeric triplet, (c0, c1, c2),}
\usage{
get_std_trip(trip, tol = 1e-10)
}
\arguments{
\item{trip}{<dbl> triplet c(c0, c1, c2)}

\item{tol}{<dbl> lower bound for norm(c1, c2)}
}
\value{
scaled homogeneous coefficients
}
\description{
of homogeneous coefficients defines a line in the
(x, y) plane.
line: c0 + c1 x + c2 y = 0
Return an equivalent scaled version of the triplet,
such that (c1, c2) is a 2-vector of unit norm, with
c2 >= 0.
}
