% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ref_materials.R
\name{get_symbol_tbl}
\alias{get_symbol_tbl}
\title{-- return a tibble of technical symbols}
\usage{
get_symbol_tbl()
}
\value{
tibble of technical symbols
}
\description{
-- return a tibble of technical symbols
}
