% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_2D.R
\name{l2D_get_xy_stats}
\alias{l2D_get_xy_stats}
\title{Given data frame df and 3 specified columns,}
\usage{
l2D_get_xy_stats(df, x_1, x_2, y_group, tol = 1e-10)
}
\arguments{
\item{df}{<df>  a data frame containing (x_1, x_2, y_group)}

\item{x_1}{<id>  name of 1st predictor variable}

\item{x_2}{<id>  name of 2nd predictor variable}

\item{y_group}{<id>  name of grouping variable}

\item{tol}{<dbl> lower bound for abs(det())}
}
\value{
list of statistics for the specified columns
}
\description{
return a list of statistics for the given columns.
}
