% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_2D.R
\name{l2D_select_xy_tbl}
\alias{l2D_select_xy_tbl}
\title{Given data frame df and 3 specified columns,}
\usage{
l2D_select_xy_tbl(df, x_1, x_2, y_group)
}
\arguments{
\item{df}{<df>  a data frame containing (x_1, x_2, y_group)}

\item{x_1}{<id>  name of 1st predictor variable}

\item{x_2}{<id>  name of 2nd predictor variable}

\item{y_group}{<id>  name of grouping variable}
}
\value{
3-column tibble extracted from input df
}
\description{
return a tibble consisting of just those columns.
}
