% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_2D.R
\name{l2D_sim_xy_tbl}
\alias{l2D_sim_xy_tbl}
\title{Randomly generate a 3-column tibble (x_1, x_2, y_group).  By}
\usage{
l2D_sim_xy_tbl(mu_mat = NULL, pi_vec = NULL, cov_mat = NULL, n_rows = 720L)
}
\arguments{
\item{mu_mat}{<dbl> means of Gaussian mixture components}

\item{pi_vec}{<dbl> 3 group probabilities (sum = 1)}

\item{cov_mat}{<dbl> conditional covariance of (Z_1, Z_2)}

\item{n_rows}{<int> desired number of rows of data}
}
\value{
random (x_1, x_2, y_group) tibble
}
\description{
default (x_1, x_2) are conditionally independent given y_group,
and are each mixtures of independent Gaussian variables.
y_group %in% 1:3
E( (X_1, X_2) | Y = k)
= sum_j pi_j * (mu_1 [[j, k]], mu_2 [[j, k]])
Code adapted from from Knox_2018_ML.
}
