% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_helpers.R
\name{list_lda_tbls}
\alias{list_lda_tbls}
\title{return a list of tibbles from an LDA_XXX object}
\usage{
list_lda_tbls(lda_xxx)
}
\arguments{
\item{lda_xxx}{<lst> output from topicmodels:LDA()}
}
\value{
list of LDA-derived tibbles
}
\description{
input:  object of class "LDA_VEM" or "LDA_Gibbs"
return: list of component tibbles
}
