% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ref_materials.R
\name{list_links}
\alias{list_links}
\title{-- reformat reference tibble as [title](link) by auth}
\usage{
list_links(x, t_idx = 1L, l_idx = 2L, a_idx = NULL)
}
\arguments{
\item{x}{<tbl> tibble containing (title, link, auth)}

\item{t_idx}{<int> index of column containing title}

\item{l_idx}{<int> index of column containing link}

\item{a_idx}{<int> index of column containing author}
}
\value{
tibble reformatted as [title](link) by auth
}
\description{
-- reformat reference tibble as [title](link) by auth
}
