% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_per_state.R
\name{list_state_arrests}
\alias{list_state_arrests}
\title{Return tibbles of (assault, rape, murder) per state.}
\usage{
list_state_arrests()
}
\value{
(long, wide) tibbles of (assault, rape, murder) per state
}
\description{
1973 violent crime stats from McNeil, 1977.
}
