% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_per_state.R
\name{list_state_stats}
\alias{list_state_stats}
\title{Join USArrests with other US state statistics}
\usage{
list_state_stats()
}
\value{
tibble: USArrests, additional US state statistics
}
\description{
Join USArrests with other US state statistics
}
