% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oecd_bli_indicators.R
\docType{data}
\name{oecd_bli_indicators}
\alias{oecd_bli_indicators}
\title{OECD Better Life Index Indicator Metadata}
\format{
A tibble with 24 rows and 5 variables:
\describe{
  \item{code}{Indicator code matching column names in \code{oecd_bli}}
  \item{category}{Dimension/topic (one of 11 well-being dimensions)}
  \item{indicator}{Indicator name}
  \item{unit}{Unit of measurement (PC = percent, YR = years,
    USD = US dollars, RATIO = ratio, HOUR = hours,
    AVSCORE = average score, MICRO_M3 = micrograms per cubic meter)}
  \item{description}{Brief description of what the indicator measures}
}
}
\source{
OECD Better Life Index,
  \url{https://www.oecdbetterlifeindex.org/}.
}
\usage{
oecd_bli_indicators
}
\description{
Descriptions of the 24 component indicators used in the
OECD Better Life Index.
}
\examples{
data(oecd_bli_indicators)
oecd_bli_indicators

}
\seealso{
\code{\link{oecd_bli}}
}
\keyword{datasets}
