% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_helpers.R
\name{para2token}
\alias{para2token}
\title{input:  a text file whose "lines" are actually paragraphs}
\usage{
para2token(
  file,
  skip_empty_rows = TRUE,
  remove_stopword = TRUE,
  strip_punct = TRUE
)
}
\arguments{
\item{file}{passed to readr::read_lines()}

\item{skip_empty_rows}{<lgl> ignore blank rows}

\item{remove_stopword}{<lgl> remove stop-words}

\item{strip_punct}{<lgl> remove punctuation}
}
\value{
list of tibbles: input, sentence, token, paragraph
}
\description{
return list of tibbles whose distinctive columns are:
- input_tbl [para]: original input paragraphs
- sentence_tbl [sntc]: sentences stripped of punctuation
- token_tbl [word]: words stripped of stop-words
- para_tbl [para]: paragraphs reconstructed from token_tbl
}
