% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_aesthetics.R
\name{pc_rot_sum}
\alias{pc_rot_sum}
\title{Repair rotation matrix returned by stats:prcomp().}
\usage{
pc_rot_sum(prcomp_obj)
}
\arguments{
\item{prcomp_obj}{<lst> object returned by stats::prcomp()}
}
\value{
principle component object, rotation matrix reoriented
}
\description{
Require sum(PC_1) > 0.
}
