% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse_per_grp.R
\name{pt_ctr_dist}
\alias{pt_ctr_dist}
\title{pt_ctr_dist}
\usage{
pt_ctr_dist(df, grp_name)
}
\arguments{
\item{df}{<df> data frame}

\item{grp_name}{<chr> name of grouping variable}
}
\value{
within-group list: mean-deviation per var, deviation-norm
}
\description{
pt_ctr_dist
}
