% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_ball.R
\name{ratio_vol_cube_ball}
\alias{ratio_vol_cube_ball}
\title{- volume of cube divided by}
\usage{
ratio_vol_cube_ball(d = 1L, check_d = FALSE)
}
\arguments{
\item{d}{<int> dimension}

\item{check_d}{<lgl> validate args}
}
\value{
volume ratio, cube / (inscribed ball)
}
\description{
- volume of inscribed ball
}
