% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_2D.R
\name{seg_to_bbox}
\alias{seg_to_bbox}
\title{Given a segment defined by a quartet of end-point coordinates}
\usage{
seg_to_bbox(seg = NULL, tol = 1e-10)
}
\arguments{
\item{seg}{<dbl> quartet c(x, y, xend, yend)}

\item{tol}{<dbl> lower bound for (x_nrm, y_nrm)}
}
\value{
bounding box encompassing input segment
}
\description{
seg = c(x, y, xend, yend)
return a bounding box, which is a list of the form
bbox = list(xmin, xmax, ymin, ymax)
Map seg end-points to opposite bbox corners.
}
