% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_2D.R
\name{seg_to_trip}
\alias{seg_to_trip}
\title{Given a segment defined by a quartet of end-point coordinates}
\usage{
seg_to_trip(seg, tol = 1e-10)
}
\arguments{
\item{seg}{<dbl> quartet c(x, y, xend, yend)}

\item{tol}{<dbl> lower bound for norm(seg)}
}
\value{
homogeneous coefficients (c0, c1, c2)
}
\description{
(x, y, xend, yend), return the line containing the segment as
a triplet, (c0, c1, c2), of homogeneous coefficients.
}
