% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_2D.R
\name{seg_valid}
\alias{seg_valid}
\title{Validate a purported directed segment of the following form.}
\usage{
seg_valid(seg, tol = 1e-10)
}
\arguments{
\item{seg}{<dbl> vector c( x, y, xend, yend )}

\item{tol}{<dbl> lower bound for norm(c1, c2)}
}
\value{
validation (T/F) of directed segment
}
\description{
seg = c( x, y, xend, yend )
}
