% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_ds_tbl.R
\name{select_ds_per_pkg}
\alias{select_ds_per_pkg}
\title{return tibble of selected data set names per package}
\usage{
select_ds_per_pkg(pkgs, datasets)
}
\arguments{
\item{pkgs}{<chr> package name(s)}

\item{datasets}{<chr> names of data sets per package}
}
\value{
selected data set names per package
}
\description{
"pkgs" notes
1: named packages are assumed to be installed
2: length must be consistent with "datasets"
}
