% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_2D.R
\name{sim_1D_xy_tbl}
\alias{sim_1D_xy_tbl}
\title{Randomly generate a 3-column tibble (x_1, i_grp, y_group) from}
\usage{
sim_1D_xy_tbl(xy_params = NULL, n_rows = 1000L)
}
\arguments{
\item{xy_params}{<tbl> prescribed mean, sd, prob, label per group}

\item{n_rows}{<int> desired number of rows of data}
}
\value{
list of input/output tibbles, plus summaries
}
\description{
mixture specifications given by input tibble xy_params.
Return input and output tibbles, along with output summaries.
}
