% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_2D.R
\name{ss_point}
\alias{ss_point}
\title{Return intersection point of 2 given line segments.}
\usage{
ss_point(seg_1, seg_2, tol = 1e-10)
}
\arguments{
\item{seg_1}{<dbl> quartet c(x, y, xend, yend)}

\item{seg_2}{<dbl> quartet c(x, y, xend, yend)}

\item{tol}{<dbl> upper bound of effective vector equality}
}
\value{
intersection point of input segments
}
\description{
Return NULL if segments do not intersect.
Each segment is defined by a quartet of
end-point coordinates (x, y, xend, yend).
}
