% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_2D.R
\name{st_point}
\alias{st_point}
\title{Determine the point (x, y) [if it exists] where a segment,}
\usage{
st_point(seg, trip)
}
\arguments{
\item{seg}{<dbl> quartet c(x, y, xend, yend)}

\item{trip}{<dbl> triplet c(c0, c1, c2)}
}
\value{
(line, segment) point of intersection
}
\description{
defined by a quartet of end-point coordinates (x, y, xend, yend),
meets a line defined by a triplet, (c0, c1, c2), of
homogeneous coefficients.
}
