% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_ball.R
\name{surf_d_ball}
\alias{surf_d_ball}
\title{- area of the surface of a d-dimensional ball of radius r}
\usage{
surf_d_ball(d = 1L, r = 1, check_d = FALSE)
}
\arguments{
\item{d}{<int> dimension}

\item{r}{<dbl> radius}

\item{check_d}{<lgl> validate args}
}
\value{
surface area of a d-dimensional ball of radius r
}
\description{
- i.e., area of a sphere of local dimension (d - 1)
}
