% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_2D.R
\name{trip_meets_seg}
\alias{trip_meets_seg}
\title{Determine whether a line defined by a triplet, (c0, c1, c2),}
\usage{
trip_meets_seg(trip, seg)
}
\arguments{
\item{trip}{<dbl> triplet c(c0, c1, c2)}

\item{seg}{<dbl> quartet c(x, y, xend, yend)}
}
\value{
validation (T/F) of (line, segment) intersection
}
\description{
of homogeneous coefficients intersects a line segment defined
by a quartet of end-point coordinates (x, y, xend, yend).
}
