% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_2D.R
\name{trip_valid}
\alias{trip_valid}
\title{Validate a purported triplet of homogeneous coefficients,}
\usage{
trip_valid(trip, tol = 1e-10)
}
\arguments{
\item{trip}{<dbl> vector c( c0, c1, c2 )}

\item{tol}{<dbl> lower bound for norm(c1, c2)}
}
\value{
validation (T/F) of homogeneous coefficients
}
\description{
which is a vector of the following form.
trip = c( c0, c1, c2 )
}
