% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_2D.R
\name{tt_point}
\alias{tt_point}
\title{Each of two lines is defined by a triplet, (c0, c1, c2),}
\usage{
tt_point(trip_1, trip_2, tol = 1e-10)
}
\arguments{
\item{trip_1}{<dbl> 1st triplet}

\item{trip_2}{<dbl> 2nd triplet}

\item{tol}{<dbl> lower bound for abs(det())}
}
\value{
(line, line) point of intersection
}
\description{
of homogeneous coefficients.  Return the point of
intersection of the two lines.
line: c0 + c1 x + c2 y = 0
}
