% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_ball.R
\name{uc_norm_stats}
\alias{uc_norm_stats}
\title{- Asymptotic (mean, variance, sd) of:}
\usage{
uc_norm_stats(d = 1L, r = 1, check_d = FALSE)
}
\arguments{
\item{d}{<int> dimension}

\item{r}{<dbl> radius}

\item{check_d}{<lgl> validate args}
}
\value{
norm of a uniform random d-vector
}
\description{
- the norm of a random d-vector X
generated by a uniform distribution
over r * the unit cube
}
