% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_2D.R
\name{valid_xy_params}
\alias{valid_xy_params}
\title{Check K-by-4 input tibble:}
\usage{
valid_xy_params(xy_params, tol = 1e-10)
}
\arguments{
\item{xy_params}{<tbl> prescribed mean, sd, prob, label per group}

\item{tol}{<dbl> lower bound to compare abs(x) to zero}
}
\value{
validation (T/F) of input tibble
}
\description{
mean, sd, prob, label per group
}
