% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_ball.R
\name{vol_d_ball}
\alias{vol_d_ball}
\title{- volume of a d-dimensional ball of radius r}
\usage{
vol_d_ball(d = 1L, r = 1, check_d = FALSE)
}
\arguments{
\item{d}{<int> dimension}

\item{r}{<dbl> radius}

\item{check_d}{<lgl> validate args}
}
\value{
volume of a d-dimensional ball of radius r
}
\description{
- volume of a d-dimensional ball of radius r
}
