% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_ball.R
\name{vol_d_cube}
\alias{vol_d_cube}
\title{- volume of a d-dimensional cube of radius r}
\usage{
vol_d_cube(d = 1L, r = 1, check_d = FALSE)
}
\arguments{
\item{d}{<int> dimension}

\item{r}{<dbl> radius}

\item{check_d}{<lgl> validate args}
}
\value{
volume of a d-dimensional cube of radius r
}
\description{
- i.e., a cube having edge-length (2 * r)
- that envelopes the tangent d-ball
}
