% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wine_quality.R
\docType{data}
\name{wine_quality}
\alias{wine_quality}
\title{Wine Quality Dataset}
\format{
A tibble with 6497 rows and 13 variables:
\describe{
  \item{color}{Wine color (factor): red or white}
  \item{fixed.acidity}{Fixed acidity in g/dm3}
  \item{volatile.acidity}{Volatile acidity in g/dm3}
  \item{citric.acid}{Citric acid in g/dm3}
  \item{residual.sugar}{Residual sugar in g/dm3}
  \item{chlorides}{Chlorides in g/dm3}
  \item{free.sulfur.dioxide}{Free sulfur dioxide in mg/dm3}
  \item{total.sulfur.dioxide}{Total sulfur dioxide in mg/dm3}
  \item{density}{Density in g/cm3}
  \item{pH}{pH value}
  \item{sulphates}{Sulphates in g/dm3}
  \item{alcohol}{Alcohol percentage by volume}
  \item{quality}{Quality score from 0 to 10}
}
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/wine+quality}
}
\usage{
wine_quality
}
\description{
Red and white wine samples from the UCI Machine Learning Repository,
with physicochemical properties and quality ratings.
}
\examples{
data(wine_quality)
summary(wine_quality)

}
\keyword{datasets}
