% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtabs_to_jaccard.R
\name{xtabs_to_jaccard}
\alias{xtabs_to_jaccard}
\title{xtabs_to_jaccard}
\usage{
xtabs_to_jaccard(xt_counts)
}
\arguments{
\item{xt_counts}{xtabs() output, named matrix of counts}
}
\value{
tibble list: xtabs output as Jaccard similarity
}
\description{
xtabs_to_jaccard
}
