% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_2D.R
\name{xy_in_bbox}
\alias{xy_in_bbox}
\title{Determine whether point(s) (x, y) belongs to}
\usage{
xy_in_bbox(x, y, bbox)
}
\arguments{
\item{x}{<dbl> 1st coordinate of point (x, y)}

\item{y}{<dbl> 2nd coordinate of point (x, y)}

\item{bbox}{<lst> named list (xmin, xmax, ymin, ymax)}
}
\value{
T/F: input points fall within bounding box
}
\description{
the closed rectangle defined by a bounding box,
which is a list of the following form.
bbox = list (xmin, xmax, ymin, ymax)
}
